+++ web/src/lib/components/timeline-viewer/timeline-day/selectable-timeline-day.svelte
@@ -7,10 +7,11 @@
   import { uploadAssetsStore } from '$lib/stores/upload';
   import { navigate } from '$lib/utils/navigation';
 
-  import AssetDateGroupComp from '$lib/components/photos-page/asset-date-group-comp.svelte';
+  import TimelineDay from '$lib/components/timeline-viewer/timeline-day/timeline-day.svelte';
   import { DayGroup } from '$lib/managers/timeline-manager/day-group.svelte';
-  import { assetsSnapshot, type Snippet } from '$lib/managers/timeline-manager/utils.svelte';
+  import { assetsSnapshot } from '$lib/managers/timeline-manager/utils.svelte';
   import { searchStore } from '$lib/stores/search.svelte';
+  import type { Snippet } from 'svelte';
 
   let { isUploading } = uploadAssetsStore;
 
@@ -22,11 +23,13 @@
     monthGroup: MonthGroup;
     timelineManager: TimelineManager;
     assetInteraction: AssetInteraction;
+
     customLayout?: Snippet<[TimelineAsset]>;
+
+    onAssetOpen?: (dayGroup: DayGroup, asset: TimelineAsset, defaultAssetOpen: () => void) => void;
     onSelect: (asset: TimelineAsset) => void;
     onScrollCompensation: (compensation: { heightDelta?: number; scrollTop?: number }) => void;
     onScrollToTop: () => void;
-    onAssetOpen?: (dayGroup: DayGroup, asset: TimelineAsset, defaultAssetOpen: () => void) => void;
   }
 
   let {
@@ -36,12 +39,12 @@
     showArchiveIcon,
     monthGroup = $bindable(),
     assetInteraction,
-    timelineManager,
     customLayout,
+    timelineManager,
+    onAssetOpen,
     onSelect,
     onScrollCompensation,
     onScrollToTop,
-    onAssetOpen,
   }: Props = $props();
 
   let lastAssetMouseEvent: TimelineAsset | null = $state(null);
@@ -259,7 +262,7 @@
 
 <svelte:document onkeydown={onKeyDown} onkeyup={onKeyUp} />
 
-<AssetDateGroupComp
+<TimelineDay
   {customLayout}
   {isSelectionMode}
   {singleSelect}
@@ -267,7 +270,6 @@
   {showArchiveIcon}
   {monthGroup}
   {timelineManager}
-  {onThumbnailClick}
   {onScrollCompensation}
   onHover={handleOnHover}
   onAssetOpen={handleOnAssetOpen}
@@ -277,4 +279,4 @@
   isAssetSelected={(asset) => assetInteraction.hasSelectedAsset(asset.id) || timelineManager.albumAssets.has(asset.id)}
   isAssetSelectionCandidate={(asset) => assetInteraction.hasSelectionCandidate(asset.id)}
   isAssetDisabled={(asset) => timelineManager.albumAssets.has(asset.id)}
-></AssetDateGroupComp>
+/>
